+++ packages/pure-parse/src/parsers/index.ts
@@ -4,8 +4,9 @@ export * from './tuples'
 export * from './object'
 export * from './oneOf'
 export * from './primitives'
-export * from './always'
 export * from './literal'
 export * from './optional'
 export * from './instanceOf'
 export * from './never'
+export * from './unknown'
+export * from './defaults'
