+++ packages/pure-parse/src/parsers/never.test.ts
@@ -1,31 +1,8 @@
 import { describe, expect, it } from 'vitest'
-import { failWith, parseNever } from './never'
-import { isSuccess, ParseFailure, Parser, ParseSuccess } from './types'
+import { parseNever } from './never'
+import { isSuccess, ParseFailure, ParseSuccess } from './types'
 import { Equals } from '../internals'
-import { InfallibleParser, UnsuccessfulParser } from './always'
 import { Infer } from '../common'
-
-describe('never', () => {
-  it('always fails', () => {
-    const parse = failWith('always fail')
-    expect(isSuccess(parse(''))).toEqual(false)
-    expect(isSuccess(parse(0))).toEqual(false)
-    expect(isSuccess(parse(false))).toEqual(false)
-    expect(isSuccess(parse(true))).toEqual(false)
-    expect(isSuccess(parse(undefined))).toEqual(false)
-    expect(isSuccess(parse(null))).toEqual(false)
-    expect(isSuccess(parse({}))).toEqual(false)
-    expect(isSuccess(parse([]))).toEqual(false)
-  })
-  describe('types', () => {
-    it('infers infallible parsers', () => {
-      const parser = failWith('never always fails')
-      const t1: Equals<typeof parser, InfallibleParser<string>> = false
-      const t2: Equals<typeof parser, Parser<string>> = false
-      const t3: Equals<typeof parser, UnsuccessfulParser> = true
-    })
-  })
-})
 
 describe('parseNever', () => {
   it('it always fails', () => {
